jenkins-api/requirements.txt
fastapi
uvicorn
requests
python-jenkins